+++ services/blockchain-indexer/shared/dataService/business/dpos/constants.js
@@ -1,6 +1,6 @@
 /*
  * LiskHQ/lisk-service
- * Copyright © 2022 Lisk Foundation
+ * Copyright © 2021 Lisk Foundation
  *
  * See the LICENSE file at the top-level directory of this distribution
  * for licensing information.
@@ -13,38 +13,14 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const { getVotes } = require('./votesSent');
-
-const {
-	WAIT_TIME_VOTER,
-	WAIT_TIME_SELF_VOTE,
-	PUNISH_TIME_VOTER,
-	PUNISH_TIME_SELF_VOTE,
-} = require('./constants');
-
-const {
-	calculatePomEndHeight,
-	standardizePomHeight,
-	findPomHeightForUnlock,
-} = require('./pom');
-
-const {
-	calculateUnlockEndHeight,
-	standardizeUnlockHeight,
-} = require('./unlock');
+const WAIT_TIME_VOTER = 2000;
+const WAIT_TIME_SELF_VOTE = 260000;
+const PUNISH_TIME_VOTER = 260000;
+const PUNISH_TIME_SELF_VOTE = 780000;
 
 module.exports = {
-	getVotes,
-
 	WAIT_TIME_VOTER,
 	WAIT_TIME_SELF_VOTE,
 	PUNISH_TIME_VOTER,
 	PUNISH_TIME_SELF_VOTE,
-
-	calculatePomEndHeight,
-	standardizePomHeight,
-
-	findPomHeightForUnlock,
-	calculateUnlockEndHeight,
-	standardizeUnlockHeight,
 };
