+++ services/blockchain-indexer/shared/dataService/business/dpos/pom.js
@@ -13,10 +13,12 @@
  * Removal or modification of this copyright notice is prohibited.
  *
  */
-const WAIT_TIME_VOTER = 2000;
-const WAIT_TIME_SELF_VOTE = 260000;
-const PUNISH_TIME_VOTER = 260000;
-const PUNISH_TIME_SELF_VOTE = 780000;
+const {
+	WAIT_TIME_VOTER,
+	WAIT_TIME_SELF_VOTE,
+	PUNISH_TIME_VOTER,
+	PUNISH_TIME_SELF_VOTE,
+} = require('./constants');
 
 const calculatePomEndHeight = pomHeight => pomHeight + PUNISH_TIME_SELF_VOTE;
 
@@ -52,42 +54,15 @@ const findPomHeightForUnlock = (unlock, account, isSelfVote) => {
 		.sort((a, b) => b - a)
 		.filter(
 			pomHeight => pomHeight < unvoteHeight + unlockWaitingPeriod
-				&& unvoteHeight < pomHeight + pomWaitingPeriod
-				&& unvoteHeight + unlockWaitingPeriod < pomHeight + pomWaitingPeriod,
+                && unvoteHeight < pomHeight + pomWaitingPeriod
+                && unvoteHeight + unlockWaitingPeriod < pomHeight + pomWaitingPeriod,
 		);
 
 	return pomHeightForUnlock || null;
 };
 
-const calculateUnlockEndHeight = (unlock, account, delegateAcc) => {
-	if (unlock.delegateAddress === account.address) { // self-unvote
-		const pomHeight = findPomHeightForUnlock(unlock, account, true);
-		return pomHeight
-			? pomHeight + PUNISH_TIME_SELF_VOTE
-			: unlock.unvoteHeight + WAIT_TIME_SELF_VOTE;
-	}
-
-	const pomHeight = findPomHeightForUnlock(unlock, delegateAcc, false);
-	return pomHeight
-		? pomHeight + PUNISH_TIME_VOTER
-		: unlock.unvoteHeight + WAIT_TIME_VOTER;
-};
-
-const standardizeUnlockHeight = (unlock, account, delegateAcc) => ({
-	start: unlock.unvoteHeight,
-	end: calculateUnlockEndHeight(unlock, account, delegateAcc),
-});
-
 module.exports = {
-	constants: {
-		WAIT_TIME_VOTER,
-		WAIT_TIME_SELF_VOTE,
-		PUNISH_TIME_VOTER,
-		PUNISH_TIME_SELF_VOTE,
-	},
 	calculatePomEndHeight,
 	standardizePomHeight,
 	findPomHeightForUnlock,
-	calculateUnlockEndHeight,
-	standardizeUnlockHeight,
 };
